Ei käytetä Redis-cachea ParameterActorissa
