+++ src/test/scala/fi/vm/sade/hakurekisteri/integration/parametrit/ParameterActorSpec.scala
@@ -22,7 +22,6 @@ class ParameterActorSpec extends ScalatraFunSuite with Matchers with AsyncAssert
 
   implicit val timeout: Timeout = 60.seconds
   val parameterConfig = ServiceConfig(serviceUrl = "http://localhost/ohjausparametrit-service")
-  val cacheFactory = MockCacheFactory.get
 
   def createEndPoint = {
     val e = mock[Endpoint]
@@ -39,7 +38,7 @@ class ParameterActorSpec extends ScalatraFunSuite with Matchers with AsyncAssert
         implicit val ec = system.dispatcher
         val endPoint = createEndPoint
         val parameterActor = system.actorOf(Props(new HttpParameterActor(
-          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))), cacheFactory)))
+          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))))))
 
         waitFuture((parameterActor ? KierrosRequest("1.2.246.562.29.32820950486")).mapTo[HakuParams])(h => {
           h.end should be (new DateTime("2014-12-31T14:07:23.213+02:00"))
@@ -56,7 +55,7 @@ class ParameterActorSpec extends ScalatraFunSuite with Matchers with AsyncAssert
         implicit val ec = system.dispatcher
         val endPoint = createEndPoint
         val parameterActor = system.actorOf(Props(new HttpParameterActor(
-          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))), cacheFactory)))
+          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))))))
 
         waitFuture((parameterActor ? KierrosRequest("1.2.246.562.29.32820950486")).mapTo[HakuParams])(h => {
           h.end should be (new DateTime("2014-12-31T14:07:23.213+02:00"))
@@ -81,7 +80,7 @@ class ParameterActorSpec extends ScalatraFunSuite with Matchers with AsyncAssert
         implicit val ec = system.dispatcher
         val endPoint = createEndPoint
         val parameterActor = system.actorOf(Props(new HttpParameterActor(
-          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))), cacheFactory)))
+          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))))))
 
         expectFailure[NoParamFoundException](parameterActor ? KierrosRequest("1.2.246.562.29.43114244536"))
 
@@ -96,8 +95,7 @@ class ParameterActorSpec extends ScalatraFunSuite with Matchers with AsyncAssert
         implicit val ec = system.dispatcher
         val endPoint = createEndPoint
         val parameterActor = system.actorOf(Props(new HttpParameterActor(
-          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))),
-          CacheFactory.apply(new OphProperties().addDefault("suoritusrekisteri.cache.redis.enabled", "false")))))
+          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))))))
 
         expectFailure[NoParamFoundException](parameterActor ? KierrosRequest("1.2.246.562.29.foobar"))
 
@@ -112,7 +110,7 @@ class ParameterActorSpec extends ScalatraFunSuite with Matchers with AsyncAssert
         implicit val ec = system.dispatcher
         val endPoint = createEndPoint
         val parameterActor = system.actorOf(Props(new HttpParameterActor(
-          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))), cacheFactory)))
+          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))))))
 
         waitFuture((parameterActor ? IsSendingEnabled("perustiedot")).mapTo[Boolean])(b => {
           b should be (true)
@@ -129,7 +127,7 @@ class ParameterActorSpec extends ScalatraFunSuite with Matchers with AsyncAssert
         implicit val ec = system.dispatcher
         val endPoint = createEndPoint
         val parameterActor = system.actorOf(Props(new HttpParameterActor(
-          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))), cacheFactory)))
+          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))))))
 
         waitFuture((parameterActor ? IsRestrictionActive("opoUpdateGraduation")).mapTo[Boolean])(b => {
           b should be (true)
@@ -146,7 +144,7 @@ class ParameterActorSpec extends ScalatraFunSuite with Matchers with AsyncAssert
         implicit val ec = system.dispatcher
         val endPoint = createEndPoint
         val parameterActor = system.actorOf(Props(new HttpParameterActor(
-          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))), cacheFactory)))
+          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))))))
 
         waitFuture((parameterActor ? IsRestrictionActive("inactiveRestriction")).mapTo[Boolean])(b => {
           b should be (false)
@@ -164,7 +162,7 @@ class ParameterActorSpec extends ScalatraFunSuite with Matchers with AsyncAssert
         implicit val ec = system.dispatcher
         val endPoint = createEndPoint
         val parameterActor = system.actorOf(Props(new HttpParameterActor(
-          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))), cacheFactory)))
+          new VirkailijaRestClient(config = parameterConfig, aClient = Some(new AsyncHttpClient(new CapturingProvider(endPoint)))))))
 
         waitFuture((parameterActor ? IsSendingEnabled("perustiedot")).mapTo[Boolean])(b => {
           b should be (true)
