+++ src/test/scala/fi/vm/sade/hakurekisteri/rest/BatchSendingClosedSpec.scala
@@ -71,7 +71,7 @@ class BatchSendingClosedSpec extends ScalatraFunSuite with MockitoSugar with Dis
   }
   val asyncProvider = new CapturingProvider(createEndpointMock)
   val client = new VirkailijaRestClient(ServiceConfig(serviceUrl = "http://localhost/ohjausparametrit-service"), aClient = Some(new AsyncHttpClient(asyncProvider)))
-  val parameterActor = system.actorOf(Props(new HttpParameterActor(client, cacheFactory)))
+  val parameterActor = system.actorOf(Props(new HttpParameterActor(client)))
   val orgsActor = system.actorOf(Props(new HttpOrganisaatioActor(client, new MockConfig, cacheFactory)))
 
   override def stop(): Unit = {
